Update the call to be one hour later in the European morning slot
This was discussed in the September 2025 meet and nobody disagreed. The
problem is that few people in the UK are able to join at 9AM their time.
And to my knowledge we havevn't had someone attending this from a timezone
in Asia which was one of the aims of the early meet.